refactor(rmcpk-core): migrates model and errors into rmcp-core
