+++ crates/rmcp-core/src/error.rs
