+++ crates/rmcp-core/src/model/annotated.rs
