+++ crates/rmcp-core/src/model/content.rs
