+++ crates/rmcp-core/src/model/extension.rs
@@ -61,7 +61,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// assert!(ext.insert(5i32).is_none());
     /// assert!(ext.insert(4u8).is_none());
@@ -79,7 +79,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// assert!(ext.get::<i32>().is_none());
     /// ext.insert(5i32);
@@ -98,7 +98,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// ext.insert(String::from("Hello"));
     /// ext.get_mut::<String>().unwrap().push_str(" World");
@@ -118,7 +118,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// *ext.get_or_insert(1i32) += 2;
     ///
@@ -134,7 +134,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// *ext.get_or_insert_with(|| 1i32) += 2;
     ///
@@ -158,7 +158,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// *ext.get_or_insert_default::<i32>() += 2;
     ///
@@ -175,7 +175,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// ext.insert(5i32);
     /// assert_eq!(ext.remove::<i32>(), Some(5i32));
@@ -193,7 +193,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// ext.insert(5i32);
     /// ext.clear();
@@ -212,7 +212,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// assert!(ext.is_empty());
     /// ext.insert(5i32);
@@ -228,7 +228,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext = Extensions::new();
     /// assert_eq!(ext.len(), 0);
     /// ext.insert(5i32);
@@ -247,7 +247,7 @@ impl Extensions {
     /// # Example
     ///
     /// ```
-    /// # use rmcp::model::Extensions;
+    /// # use rmcp_core::model::Extensions;
     /// let mut ext_a = Extensions::new();
     /// ext_a.insert(8u8);
     /// ext_a.insert(16u16);
