+++ crates/rmcp-core/src/model/meta.rs
