+++ crates/rmcp-core/src/model/prompt.rs
