+++ crates/rmcp-core/src/model/resource.rs
