+++ crates/rmcp-core/src/model/serde_impl.rs
