+++ crates/rmcp-core/src/model/tool.rs
