+++ crates/rmcp/src/handler/client.rs
@@ -1,5 +1,5 @@
 use crate::{
-    error::Error as McpError,
+    Error as McpError,
     model::*,
     service::{Peer, RequestContext, RoleClient, Service, ServiceRole},
 };
