+++ examples/clients/Cargo.toml
@@ -1,5 +1,3 @@
-
-
 [package]
 name = "mcp-client-examples"
 version = "0.1.5"
@@ -13,12 +11,13 @@ rmcp = { path = "../../crates/rmcp", features = [
     "transport-child-process",
     "tower"
 ] }
+rmcp-core = { path = "../../crates/rmcp-core", features = ["macros"] }
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-rand = "0.8"
+rand = "0.9"
 futures = "0.3"
 anyhow = "1.0"
