+++ examples/servers/Cargo.toml
@@ -1,5 +1,3 @@
-
-
 [package]
 name = "mcp-server-examples"
 version = "0.1.5"
@@ -7,7 +5,8 @@ edition = "2024"
 publish = false
 
 [dependencies]
-rmcp= { path = "../../crates/rmcp", features = ["server", "transport-sse-server", "transport-io"] }
+rmcp= { path = "../../crates/rmcp", features = ["server", "transport-sse-server", "transport-io", "macros"] }
+rmcp-core = { path = "../../crates/rmcp-core" }
 tokio = { version = "1", features = ["macros", "rt", "rt-multi-thread", "io-std", "signal"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
